+++ routes/api.php
@@ -1,6 +1,1 @@
 <?php
-
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
-
-
