+++ routes/custom/admin_routes.php
@@ -10,19 +10,12 @@ use App\Http\Controllers\Admin\UserRoleController;
 use App\Http\Controllers\Admin\VisaTypeController;
 use Illuminate\Support\Facades\Route;
 
-
-//Auth Routes
 Route::group([
     'as' => 'admin.',
-    'prefix' => 'admin'
-], function () {
-    Route::post('signin', [AdminAuthController::class, 'signin'])->name('signin');
-});
-
-Route::group([
-    'as' => 'admin.',
-    'prefix' => 'admin',
-    'middleware' => ['auth:api']
+    'prefix' => 'admin-panel',
+    'middleware' => [
+        'auth:api'
+    ]
 ], function () {
     Route::apiResource('permission-groups', PermissionGroupController::class);
     Route::apiResource('countries', CountryController::class)->only(['index', 'show', 'update']);
@@ -31,5 +24,4 @@ Route::group([
     Route::apiResource('visa-types', VisaTypeController::class);
     Route::apiResource('roles', RoleController::class);
     Route::apiResource('users.roles', UserRoleController::class)->only(['update', 'destroy']);
-    Route::post('signout', [AdminAuthController::class, 'signout'])->name('signout');
 });
