+++ routes/custom/website_routes.php
@@ -10,11 +10,10 @@ use Illuminate\Support\Facades\Route;
 
 Route::group([
     'as' => 'website.',
-    'middleware' => ['auth:api']
+    'middleware' => [
+        'auth:api'
+    ]
 ], function () {
-    Route::post('signout', [WebsiteAuthController::class, 'signout']);
-    Route::post('change-password', [WebsiteAuthController::class, 'changePassword']);
-    Route::post('verify-email', [WebsiteAuthController::class, 'verifyEmail']);
     Route::get('profile', [UserProfileController::class, 'index'])->name('profile.index');
     Route::put('profile', [UserProfileController::class, 'update'])->name('profile.update');
     Route::delete('profile', [UserProfileController::class, 'destroy'])->name('profile.destroy');
@@ -26,6 +25,4 @@ Route::group([
     Route::apiResource('countries', CountryController::class)->only(['index', 'show']);
     Route::apiResource('employment-types', EmploymentTypeController::class)->only(['index', 'show']);
     Route::apiResource('visa-types', VisaTypeController::class)->only(['index', 'show']);
-    Route::post('signup', [WebsiteAuthController::class, 'signup']);
-    Route::post('signin', [WebsiteAuthController::class, 'signin']);
 });
